+++ Public/Assistants/New-Assistant.ps1
@@ -28,6 +28,7 @@ function New-Assistant {
             'gpt-4-32k-0613',
             'gpt-4-turbo',
             'gpt-4-turbo-2024-04-09',
+            'gpt-4.5-preview',
             'o1',
             'o3-mini'
         )]