+++ Public/Assistants/Set-Assistant.ps1
@@ -29,6 +29,7 @@ function Set-Assistant {
             'gpt-4-32k-0613',
             'gpt-4-turbo',
             'gpt-4-turbo-2024-04-09',
+            'gpt-4.5-preview',
             'o1',
             'o3-mini'
         )]