+++ Public/Chat/Request-ChatCompletion.ps1
@@ -34,6 +34,7 @@ function Request-ChatCompletion {
             'gpt-4-32k-0613',
             'gpt-4-turbo',
             'gpt-4-turbo-2024-04-09',
+            'gpt-4.5-preview',
             'o1',
             'o1-mini',
             'o1-preview',