+++ Public/Runs/Start-ThreadRun.ps1
@@ -31,6 +31,7 @@ function Start-ThreadRun {
             'gpt-4-32k-0613',
             'gpt-4-turbo',
             'gpt-4-turbo-2024-04-09',
+            'gpt-4.5-preview',
             'o1',
             'o3-mini'
         )]